+++ 2406/2/src/Models/artist.rb
