+++ 2406/2/src/Models/battle.rb
